+++ src/Kusto/test/Stop-AzKustoCluster.Tests.ps1
@@ -5,7 +5,7 @@ if (-Not (Test-Path -Path $loadEnvPath)) {
 . ($loadEnvPath)
 $TestRecordingFile = Join-Path $PSScriptRoot 'Stop-AzKustoCluster.Recording.json'
 $currentPath = $PSScriptRoot
-while(-not $mockingPath) {
+while (-not $mockingPath) {
     $mockingPath = Get-ChildItem -Path $currentPath -Recurse -Include 'HttpPipelineMocking.ps1' -File
     $currentPath = Split-Path -Path $currentPath -Parent
 }
@@ -14,17 +14,19 @@ while(-not $mockingPath) {
 Describe 'Stop-AzKustoCluster' {
     It 'Stop' {
         $resourceGroupName = $env.resourceGroupName
-        $clusterName = $env.clusterName + "a"
+        $clusterName = $env.PlainClusterName
 
         { Stop-AzKustoCluster -ResourceGroupName $resourceGroupName -Name $clusterName } | Should -Not -Throw
+        Start-AzKustoCluster -ResourceGroupName $resourceGroupName -Name $clusterName
     }
 
     It 'StopViaIdentity' {
         $resourceGroupName = $env.resourceGroupName
-        $clusterName = $env.clusterName + "a"
+        $clusterName = $env.PlainClusterName
 
         $clusterGetItem = Get-AzKustoCluster -ResourceGroupName $resourceGroupName -Name $clusterName
 
         { Stop-AzKustoCluster -InputObject $clusterGetItem } | Should -Not -Throw
+        Start-AzKustoCluster -InputObject $clusterGetItem
     }
 }