+++ src/SignalR/SignalR/Cmdlets/NewAzureRmSignalR.cs
@@ -12,7 +12,6 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
-using System.Collections;
 using System.Collections.Generic;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.Common.Strategies;
@@ -21,7 +20,6 @@ using Microsoft.Azure.Commands.SignalR.Models;
 using Microsoft.Azure.Management.Internal.Resources;
 using Microsoft.Azure.Management.SignalR;
 using Microsoft.Azure.Management.SignalR.Models;
-using Microsoft.WindowsAzure.Commands.Common;
 using Newtonsoft.Json;
 
 namespace Microsoft.Azure.Commands.SignalR.Cmdlets
@@ -120,14 +118,15 @@ namespace Microsoft.Azure.Commands.SignalR.Cmdlets
 
                     Sku = Sku ?? DefaultSku;
 
-                    IList<SignalRFeature> features = ServiceMode == null ? null : new List<SignalRFeature> { new SignalRFeature(value: ServiceMode) };
+                    IList<SignalRFeature> features = ServiceMode == null ? null : new List<SignalRFeature> { new SignalRFeature(flag: FeatureFlags.ServiceMode, value: ServiceMode) };
                     SignalRCorsSettings cors = AllowedOrigin == null ? null : new SignalRCorsSettings(allowedOrigins: origins);
 
-                    var parameters = new SignalRCreateParameters(
+                    var parameters = new SignalRResource(
                         location: Location,
                         tags: Tag,
                         sku: new ResourceSku(name: Sku, capacity: UnitCount),
-                        properties: new SignalRCreateOrUpdateProperties(features: features, cors: cors));
+                        features: features,
+                        cors: cors);
 
                     Client.SignalR.CreateOrUpdate(ResourceGroupName, Name, parameters);