+++ src/Storage/Storage.Management.Test/ScenarioTests/StorageBlobTests.ps1
@@ -434,6 +434,169 @@ function Test-StorageBlobServiceProperties
     }
 }
 
+<#
+.SYNOPSIS
+Test StorageAccount Blob Restore
+.DESCRIPTION
+SmokeTest
+#>
+function Test-StorageBlobRestore
+{
+    # Setup
+    $rgname = Get-StorageManagementTestResourceName;
 
+    try
+    {
+        # Test
+        $stoname = 'sto' + $rgname;
+        $stotype = 'Standard_LRS';
+        $loc = Get-ProviderLocation ResourceManagement;
+        $kind = 'StorageV2'
+	
+        Write-Verbose "RGName: $rgname | Loc: $loc"
+        New-AzResourceGroup -Name $rgname -Location $loc;
+		
+        $loc = Get-ProviderLocation_Stage ResourceManagement;
+        New-AzStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype -Kind $kind 
+        $stos = Get-AzStorageAccount -ResourceGroupName $rgname;
+		
+        # Enable Blob Delete Retension Policy, Enable Changefeed, then enabled blob restore policy, then get blob service proeprties and check the setting
+        Enable-AzStorageBlobDeleteRetentionPolicy -ResourceGroupName $rgname -StorageAccountName $stoname -RetentionDays 5
+        Update-AzStorageBlobServiceProperty -ResourceGroupName $rgname -StorageAccountName $stoname -EnableChangeFeed $true -IsVersioningEnabled $true
+        # If record, need sleep before enable the blob restore policy, or will get server error
+        #sleep 100 
+        Enable-AzStorageBlobRestorePolicy -ResourceGroupName $rgname -StorageAccountName $stoname -RestoreDays 4
+        $property = Get-AzStorageBlobServiceProperty -ResourceGroupName $rgname -StorageAccountName $stoname
+        #Assert-AreEqual $true $property.ChangeFeed.Enabled
+        Assert-AreEqual $true $property.DeleteRetentionPolicy.Enabled
+        Assert-AreEqual 5 $property.DeleteRetentionPolicy.Days
+        Assert-AreEqual $true $property.RestorePolicy.Enabled
+        Assert-AreEqual 4 $property.RestorePolicy.Days
+
+        # restore blobs by -asjob
+        $range1 = New-AzStorageBlobRangeToRestore -StartRange container1/blob1 -EndRange container2/blob2
+        $range2 = New-AzStorageBlobRangeToRestore -StartRange container3/blob3 -EndRange ""
+        sleep 2
+        $job = Restore-AzStorageBlobRange -ResourceGroupName $rgname -StorageAccountName $stoname -TimeToRestore (Get-Date).AddSeconds(-1) -BlobRestoreRange $range1,$range2 -WaitForComplete -asjob
+
+        # Get  Storage Account with Blob Restore Status
+        $stos = Get-AzStorageAccount -ResourceGroupName $rgname -StorageAccountName $stoname -IncludeBlobRestoreStatus
+
+        # wait for restore job finish, and check Blob Restore Status in Storage Account	
+        $job | Wait-Job
+        $stos = Get-AzStorageAccount -ResourceGroupName $rgname -StorageAccountName $stoname -IncludeBlobRestoreStatus
+        Assert-AreEqual "Complete" $stos.BlobRestoreStatus.Status
+
+        Remove-AzStorageAccount -Force -ResourceGroupName $rgname -Name $stoname;
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $rgname
+    }
+}
+
+<#
+.SYNOPSIS
+Test StorageAccount Object Replication
+.DESCRIPTION
+SmokeTest
+#>
+function Test-StorageBlobORS
+{
+    # Setup
+    $rgname = Get-StorageManagementTestResourceName;
+
+    try
+    {
+        # Test
+        $stoname1 = 'sto' + $rgname + 'src';
+        $stoname2 = 'sto' + $rgname + 'dest';
+        $stotype = 'Standard_LRS';
+        $loc = Get-ProviderLocation ResourceManagement;
+        $kind = 'StorageV2'
+	
+        Write-Verbose "RGName: $rgname | Loc: $loc"
+        New-AzResourceGroup -Name $rgname -Location $loc;
+		
+        $loc = Get-ProviderLocation_Canary ResourceManagement;
+        New-AzStorageAccount -ResourceGroupName $rgname -Name $stoname1 -Location $loc -Type $stotype -Kind $kind 
+        New-AzStorageAccount -ResourceGroupName $rgname -Name $stoname2 -Location $loc -Type $stotype -Kind $kind 
+		
+		# Enable Blob Enable Changefeed and versioning
+		Update-AzStorageBlobServiceProperty -ResourceGroupName $rgname -StorageAccountName $stoname1 -EnableChangeFeed $true -IsVersioningEnabled $true
+		Update-AzStorageBlobServiceProperty -ResourceGroupName $rgname -StorageAccountName $stoname2 -EnableChangeFeed $true -IsVersioningEnabled $true
+		$property1 = Get-AzStorageBlobServiceProperty -ResourceGroupName $rgname -StorageAccountName $stoname1
+		Assert-AreEqual $true $property1.ChangeFeed.Enabled
+		Assert-AreEqual $true $property1.IsVersioningEnabled 
+		$property2 = Get-AzStorageBlobServiceProperty -ResourceGroupName $rgname -StorageAccountName $stoname2
+		Assert-AreEqual $true $property2.ChangeFeed.Enabled
+		Assert-AreEqual $true $property2.IsVersioningEnabled 
+
+		# create containers		
+		Get-AzStorageAccount -ResourceGroupName $rgname -StorageAccountName $stoname1  | New-AzRmStorageContainer -name src
+		Get-AzStorageAccount -ResourceGroupName $rgname -StorageAccountName $stoname2 | New-AzRmStorageContainer -name dest
+		Get-AzStorageAccount -ResourceGroupName $rgname -StorageAccountName $stoname1  | New-AzRmStorageContainer -name src1
+		Get-AzStorageAccount -ResourceGroupName $rgname -StorageAccountName $stoname2 | New-AzRmStorageContainer -name dest1
+
+		# create rules
+		$minCreationTime = "2019-01-01T16:00:00Z"
+		$rule1 = New-AzStorageObjectReplicationPolicyRule -SourceContainer src1 -DestinationContainer dest1 
+		$rule2 = New-AzStorageObjectReplicationPolicyRule -SourceContainer src -DestinationContainer dest -MinCreationTime $minCreationTime -PrefixMatch a,abc,dd #-Tag t1,t2,t3 
+
+		# set policy to dest account
+		$destPolicy = Set-AzStorageObjectReplicationPolicy -ResourceGroupName $rgname -StorageAccountName $stoname2 -PolicyId default -SourceAccount $stoname1  -Rule $rule1,$rule2
+		$policyID = $destPolicy.PolicyId
+		Assert-AreEqual $stoname1 $destPolicy.SourceAccount
+		Assert-AreEqual $stoname2 $destPolicy.DestinationAccount
+		Assert-AreEqual 2 $destPolicy.Rules.Count
+		Assert-AreEqual src1 $destPolicy.Rules[0].SourceContainer
+		Assert-AreEqual dest1 $destPolicy.Rules[0].DestinationContainer
+		Assert-AreEqual $null $destPolicy.Rules[0].Filters
+		Assert-AreEqual src $destPolicy.Rules[1].SourceContainer
+		Assert-AreEqual dest $destPolicy.Rules[1].DestinationContainer
+		Assert-AreEqual 3 $destPolicy.Rules[1].Filters.PrefixMatch.Count
+		Assert-AreEqual $minCreationTime ($destPolicy.Rules[1].Filters.MinCreationTime.ToUniversalTime().ToString("s")+"Z")
+		$destPolicy = Get-AzStorageObjectReplicationPolicy -ResourceGroupName $rgname -StorageAccountName $stoname2 -PolicyId $destPolicy.PolicyId
+		Assert-AreEqual $policyID $destPolicy.PolicyId
+		Assert-AreEqual $stoname1 $destPolicy.SourceAccount
+		Assert-AreEqual $stoname2 $destPolicy.DestinationAccount
+		Assert-AreEqual 2 $destPolicy.Rules.Count
+		Assert-AreEqual src1 $destPolicy.Rules[0].SourceContainer
+		Assert-AreEqual dest1 $destPolicy.Rules[0].DestinationContainer
+		Assert-AreEqual $null $destPolicy.Rules[0].Filters
+		Assert-AreEqual src $destPolicy.Rules[1].SourceContainer
+		Assert-AreEqual dest $destPolicy.Rules[1].DestinationContainer
+		Assert-AreEqual 3 $destPolicy.Rules[1].Filters.PrefixMatch.Count
+		Assert-AreEqual $minCreationTime ($destPolicy.Rules[1].Filters.MinCreationTime.ToUniversalTime().ToString("s")+"Z")
+
+		#Set policy to source account
+		Set-AzStorageObjectReplicationPolicy -ResourceGroupName $rgname -StorageAccountName $stoname1 -InputObject $destPolicy
+		$srcPolicy = Get-AzStorageObjectReplicationPolicy -ResourceGroupName $rgname -StorageAccountName $stoname1
+		Assert-AreEqual $policyID $srcPolicy.PolicyId
+		Assert-AreEqual $stoname1 $srcPolicy.SourceAccount
+		Assert-AreEqual $stoname2 $srcPolicy.DestinationAccount
+		Assert-AreEqual 2 $srcPolicy.Rules.Count
+		Assert-AreEqual src1 $srcPolicy.Rules[0].SourceContainer
+		Assert-AreEqual dest1 $srcPolicy.Rules[0].DestinationContainer
+		Assert-AreEqual $null $srcPolicy.Rules[0].Filters
+		Assert-AreEqual src $srcPolicy.Rules[1].SourceContainer
+		Assert-AreEqual dest $srcPolicy.Rules[1].DestinationContainer
+		Assert-AreEqual 3 $srcPolicy.Rules[1].Filters.PrefixMatch.Count
+		Assert-AreEqual $minCreationTime ($srcPolicy.Rules[1].Filters.MinCreationTime.ToUniversalTime().ToString("s")+"Z")
+
+		#remove policies		
+		Remove-AzStorageObjectReplicationPolicy -ResourceGroupName $rgname -StorageAccountName $stoname2 -PolicyId $destPolicy.PolicyId
+		Remove-AzStorageObjectReplicationPolicy -ResourceGroupName $rgname -StorageAccountName $stoname1 -PolicyId $srcPolicy.PolicyId
+		
+        Remove-AzStorageAccount -Force -ResourceGroupName $rgname -Name $stoname1;
+        Remove-AzStorageAccount -Force -ResourceGroupName $rgname -Name $stoname2;
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $rgname
+    }
+}