+++ tools/Gen2Master/MoveFromGeneration2Master.ps1
@@ -40,7 +40,7 @@ Function Move-Generation2Master {
         Write-Host "Copying docs: $SourceItem." -ForegroundColor Yellow
         Copy-Item -Recurse -Path $SourceItem -Destination $DestItem
         #EndRegion
-        $File2Copy = @('*.ps1', 'how-to.md', 'readme.md', '*.psm1', '*.ps1xml', 'MSSharedLibKey.snk')
+        $File2Copy = @('*.ps1', 'how-to.md', 'readme.md', '*.psm1', '*.ps1xml', 'MSSharedLibKey.snk', 'generate-info.json')
         Foreach($File in $File2Copy) {
             $SourceItem = Join-Path -Path $SourcePath -ChildPath $File
             $DestItem = Join-Path -Path $DestPath -ChildPath $File
@@ -66,9 +66,12 @@ Function Move-Generation2Master {
         If ($Null -ne $ModuleGuid) {
             $Psd1Metadata.GUID = $ModuleGuid
         }
-        If ($Null -ne $RequiredModule) {
-            $Psd1Metadata.RequiredModules = $RequiredModule
+        If ($Null -eq $RequiredModule) {
+            $AccountsModulePath = [System.IO.Path]::Combine($DestPath, '..', 'Accounts', 'Accounts')
+            $AccountsMetadata = Import-LocalizedData -BaseDirectory $AccountsModulePath -FileName "Az.Accounts.psd1"
+            $RequiredModule = @(@{ModuleName = 'Az.Accounts'; ModuleVersion = $AccountsMetadata.ModuleVersion; })
         }
+        $Psd1Metadata.RequiredModules = $RequiredModule
         If ($Psd1Metadata.FunctionsToExport -Contains "*") {
             $Psd1Metadata.FunctionsToExport = ($Psd1Metadata.FunctionsToExport | Where-Object {$_ -ne "*"})
         }
@@ -95,6 +98,13 @@ Function Move-Generation2Master {
 
         #Region update azure-powershell-modules.md
         
+        #EndRegion
+
+        #Region update GeneratedModuleList
+        $GeneratedModuleListPath = [System.IO.Path]::Combine(@($PSScriptRoot, "..", "GeneratedModuleList.txt"))
+        $Modules = Get-Content $GeneratedModuleListPath + "Az.$ModuleName"
+        $NewModules = $Modules | Sort-Object
+        Set-Content -Path $GeneratedModuleListPath -Value $NewModules
         #EndRegion
 
         Copy-Template -SourceName Az.ModuleName.csproj -DestPath $DestPath -DestName "Az.$ModuleName.csproj"