+++ src/Databricks/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* General availability of 'Az.Functions' module
+* Added support for virtual network peering
 
 ## Version 0.2.0
 * Added support for DBFS Double Encryption Support