+++ src/ManagedServices/ManagedServices/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated breaking change warnings on cmdlets of managed services assignment and definition
+
+## Version 1.1.0
+* Added breaking change warnings on cmdlets of managed services assignment and definition
 
 ## Version 1.0.2
 * Update references in .psd1 to use relative path