+++ src/Websites/Websites/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added support for Premium V3 pricing tier
 
 ## Version 1.11.0
 * Added support to perform operations for Slots not in the same resource group as the App Service Plan